backend/requirements.txt

# Auto-generated requirements
fastapi
uvicorn
python-dotenv
pandas
numpy
scikit-learn
joblib
qiskit
qiskit-aer
requests
# Image processing
scikit-image
pillow
